refactor: applying separation of concerns for desktop and mobile versions
