+++ src/components/layouts/sidebar/components/mobile/index.ts
@@ -1,0 +1,1 @@
+export * from './mobile-profile-bottom-sheet'
