+++ src/components/layouts/sidebar/components/mobile/mobile-profile-bottom-sheet.tsx
@@ -1,0 +1,9 @@
+import { Box, Text } from '@chakra-ui/react'
+
+export const MobileProfileBottomSheet = () => {
+  return (
+    <Box>
+      <Text>Mobile Profile Bottom Sheet</Text>
+    </Box>
+  )
+}
